% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Codes_09.R
\docType{data}
\name{Codes_09}
\alias{Codes_09}
\title{NHANES 2009/2010 Food codes}
\format{DRXFDCD
   Food code

DRXFCSD
   Short food code description

DRXFCLD
   Long food code description}
\usage{
Codes_09
}
\description{
An .rda data file containing 3 columns of NHANES food codes with
the original NHANES column names.  Not currently used in biteR.
}
\keyword{internal}

