% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Day1_11.R
\docType{data}
\name{Day1_11}
\alias{Day1_11}
\title{NHANES 2011/2012 Day 1 dietary survey records}
\format{SEQN
   The NHANES participant ID

WTDR2D
   The NHANES participant's 2-day dietary sample weight

DR1IFDCD
    The food code consumed.

DR1IGRMS
    Grams of the food consumed.

DR1DRSTZ
    Dietary recall status, where 1 = Reliable and met the minimum criteria.

DR1_030Z
    Name of eating occasion}
\usage{
Day1_11
}
\description{
An .rda data file containing 5 columns of NHANES dietary intake records and retaining
the original NHANES column names.
}
\keyword{internal}

