% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Day2_09.R
\docType{data}
\name{Day2_09}
\alias{Day2_09}
\title{NHANES 2009/2010 Day 2 dietary survey records}
\format{SEQN
   The NHANES participant ID

WTDR2D
   The NHANES participant's 2-day dietary sample weight

DR2IFDCD
    The food code consumed.

DR2IGRMS
    Grams of the food consumed.

DR2DRSTZ
    Dietary recall status, where 1 = Reliable and met the minimum criteria.

DR2_030Z
    Name of eating occasion}
\usage{
Day2_09
}
\description{
An .rda data file containing 5 columns of NHANES dietary intake records and retaining
the original NHANES column names.
}
\keyword{internal}

