% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Demo_09.R
\docType{data}
\name{Demo_09}
\alias{Demo_09}
\title{NHANES 2009/2010 Demographic data}
\format{SEQN
   NHANES participant ID

RIAGENDR
   Gender. 1=Male 2=Female

RIDAGEYR
   Age of NHANES participant at screening.  Individuals 80 and over are
   topcoded at 80 years.

SDMVPSU
   Masked variance pseudo-PSU

SDMVSTRA
   Masked variance pseudo-stratum}
\usage{
Demo_09
}
\description{
An .rda data file containing 5 columns of NHANES demogrpahic data with
the original NHANES column names.
}
\keyword{internal}

