% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/param.bite.R
\name{param.bite}
\alias{param.bite}
\title{Define Parameters for body intake estimation (BITE)}
\usage{
param.bite(file_name = "", type = "")
}
\arguments{
\item{file_name}{Name of input file containing additive uses and use level by food code.
See help >package?biteR for input file requirements.

type "amt" vs. "size" only determines here whether records are kept for participants
who did not complete both days of the survey. For size, as we are not concerned
with weighting, we do not insist upon it.}
}
\value{
A BITE object, instantiated with the attributes required to calculate the raw
daily intake of the additive for each NHANES participant, using raw.bite, to be passed
into raw.bite(x).
}
\description{
Defines parameters with which to build the BITE object.
}
\details{
This function uses the input file to parameterize the BITE object, primarily the
records attribute, which is a list of data.tables, one for each use and one for the
total use, containing all consumption records for all food codes in the use.
}

