% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pop.bite.R
\name{pop.bite}
\alias{pop.bite}
\title{Calculate Body Intake Population Statistics}
\usage{
pop.bite(x)
}
\arguments{
\item{x}{An object of class "bite" processed by raw.bite, and containing the demo
attribute with each NHANES participant's raw average daily consumption of the additive
by use and by total.}
}
\value{
A list of 3 matrices of population statistics in the supported demographic strata
(see help >package?biteR for supported strata).  The matrix for the two genders combined
is called "full", the female matrix is called "female"; the male matrix is called "male".
}
\description{
Returns matrices of population additive consumption or portion size.
}
\details{
When the type is 'amt', calculates population (NHANES) survey weighted statistics for an
additive by demographic strata.
When the type is 'size', calculates raw percentiles of portion size by occasion, broken
out by age only.
}

