% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/raw.bite.R
\name{raw.bite}
\alias{raw.bite}
\title{Calculate Raw Body Intake}
\usage{
raw.bite(x)
}
\arguments{
\item{x}{An object of class "bite" built by param.bite}
}
\value{
Updated bite object, with updated attribute 'demo' now including a calculated
amount of additive consumed on average per day for each NHANES dietary survey participant,
by use and total.
}
\description{
Calculates raw intake of additive for each NHANES participant.
}
\details{
This function uses the information on the parameterized BITE object to calculate
raw intake values by use and by total for every person who participated in the
NHANES dietary survey.
}

