% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Weight_11.R
\docType{data}
\name{Weight_11}
\alias{Weight_11}
\title{NHANES 2011/2012 Body weight data}
\format{SEQN
   NHANES participant ID

BMXWT
   Body weight in kilos

BMIWT
   Body weight comment; mostly missing and NA as saved by R}
\usage{
Weight_11
}
\description{
An .rda data file containing 3 columns of NHANES body weight data with
the original NHANES column names.
}
\keyword{internal}

